venvStore/requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.2
django-allauth==0.41.0
django-bootstrap4==1.1.1
django-js-asset==1.2.2
django-mptt==0.10.0
idna==2.8
oauthlib==3.1.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
